modal sin handleAccept
