+++ src/components/Modal/types.ts
@@ -2,7 +2,7 @@ import { PRIMARY, SECONDARY } from "../../common/styles/constants";
 
 export interface IModal {
     children: React.ReactNode;
-    handleAccept?: React.MouseEventHandler<HTMLButtonElement>;
+    // handleAccept: React.MouseEventHandler<HTMLButtonElement>;
     show?: boolean;
     setShow?: React.Dispatch<React.SetStateAction<boolean>>;
     styles?: string;
